+++ ui/src/pages/namespace/Explorer/Page/poc/PageCompiler/primitives/Variable/utils/useVariableJSX.ts
@@ -1,6 +1,6 @@
 import { UseVariableFailure, useVariable } from "./useVariable";
 
-import { VariableType } from "../../../../../../schema/primitives/variable";
+import { VariableType } from "../../../../schema/primitives/variable";
 import { z } from "zod";
 
 export const JSXValueSchema = z.union([
